아이템 정산
-작업자-

-한 일-
아이템 정산 ui
아이템 정산 기능
원피스 수정

+++ Assets/@Script/ScriptableObj/ItemData.cs
@@ -12,7 +12,7 @@ public class ItemData : ScriptableObject
     public float atkTime;
     public float damage;
 
-    public float money;
+    public int money;
     public int itemWeight;
     public Sprite sprite;
 }
